requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai-whisper==20231117
langchain==0.0.350
pydantic==2.5.2
python-jose==3.3.0
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
python-sox==1.4.1
pytest-asyncio==0.21.1
streamlit==1.29.0
requests==2.31.0
pydub==0.25.1
streamlit-webrtc==0.47.1
pytest-cov==4.1.0
openai==1.3.5
anthropic==0.7.4 